requirements.txt
flask 
flask-cors 
yt-dlp 
pypdf 
python-docx